working new signal diagnostics

+++ ancsim/diagnostics/diagnosticplots.py
@@ -8,30 +8,31 @@ import ancsim.experiment.multiexperimentutils as meu
 import ancsim.utilities as util
 
 
-class SCALINGTYPE(Enum):
-    linear = 1
-    dbPower = 2
-    dbAmp = 3
-    ln = 4
+# class SCALINGTYPE(Enum):
+#     linear = 1
+#     dbPower = 2
+#     dbAmp = 3
+#     ln = 4
 
 
-def scaleData(scalingType, data):
-    if scalingType == SCALINGTYPE.linear:
-        return data
-    elif scalingType == SCALINGTYPE.dbPower:
-        return 10 * np.log10(data)
-    elif scalingType == SCALINGTYPE.dbAmp:
-        return 20 * np.log10(data)
-    elif scalingType == SCALINGTYPE.ln:
-        return np.log(data)
+# def scaleData(scalingType, data):
+#     if scalingType == SCALINGTYPE.linear:
+#         return data
+#     elif scalingType == SCALINGTYPE.dbPower:
+#         return 10 * np.log10(data)
+#     elif scalingType == SCALINGTYPE.dbAmp:
+#         return 20 * np.log10(data)
+#     elif scalingType == SCALINGTYPE.ln:
+#         return np.log(data)
 
 
-def functionOfTimePlot(name, outputs, metadata, timeIdx, folder, printMethod="pdf"):
+def functionOfTimePlot(name, diags, timeIdx, folder, printMethod="pdf"):
     fig, ax = plt.subplots(1, 1, figsize=(14, 8))
     fig.tight_layout(pad=4)
 
     xValues = np.arange(timeIdx)[None, :]
-    for algoName, output in outputs.items():
+    for proc_name, diag in diags.items():
+        output = diag.get_output()
         if output.ndim == 1:
             output = output[None, :timeIdx]
         elif output.ndim == 2:
@@ -39,7 +40,7 @@ def functionOfTimePlot(name, outputs, metadata, timeIdx, folder, printMethod="pd
         else:
             raise NotImplementedError
 
-        numChannels = output.shape[0]
+        num_channels = output.shape[0]
 
         filterArray = np.logical_not(np.isnan(output))
         assert np.isclose(filterArray, filterArray[0, :]).all()
@@ -47,9 +48,9 @@ def functionOfTimePlot(name, outputs, metadata, timeIdx, folder, printMethod="pd
 
         if "label_suffix_channel" in metadata:
             labels = ["_".join((algoName, suf)) for suf in metadata["label_suffix_channel"]]
-            assert len(labels) == numChannels
+            assert len(labels) == num_channels
         else:
-            labels = [algoName for _ in range(numChannels)]
+            labels = [algoName for _ in range(num_channels)]
         ax = plotMultipleChannels(ax, xValues[:,filterArray], output[:,filterArray], labels)
 
     ax.spines["right"].set_visible(False)
@@ -84,17 +85,19 @@ def legendWithoutDuplicates(ax, loc):
     ax.legend(newHandles, newLabels, loc=loc)
 
 
-def savenpz(name, outputs, metadata, timeIdx, folder, printMethod="pdf"):
+def savenpz(name, diags, timeIdx, folder, printMethod="pdf"):
     """Keeps only the latest save.
     Assumes that the data in previous
     saves is present in the current data"""
-    flatOutputs = util.flattenDict(outputs, sep="~")
-    np.savez_compressed(folder.joinpath(name + "_" + str(timeIdx)), **flatOutputs)
+    outputs = {proc_name: diag.get_output() for proc_name, diag in diags.items()}
+    #flatOutputs = util.flattenDict(outputs, sep="~")
+    np.savez_compressed(folder.joinpath(name + "_" + str(timeIdx)), **outputs)
 
-    earlierFiles = meu.findAllEarlierFiles(folder, name, timeIdx, nameIncludesIdx=False)
-    for f in earlierFiles:
-        if f.suffix == ".npz":
-            f.unlink()
+    if list(diags.values())[0].keep_only_last_export:
+        earlierFiles = meu.findAllEarlierFiles(folder, name, timeIdx, nameIncludesIdx=False)
+        for f in earlierFiles:
+            if f.suffix == ".npz":
+                f.unlink()
 
 
 def soundfieldPlot(name, outputs, metadata, timeIdx, folder, printMethod="pdf"):
@@ -102,10 +105,10 @@ def soundfieldPlot(name, outputs, metadata, timeIdx, folder, printMethod="pdf"):
 	
 
 
-def plotIR(name, outputs, metadata, timeIdx, folder, printMethod="pdf"):
+def plotIR(name, diags, timeIdx, folder, printMethod="pdf"):
     numSets = 0
-    for algoName, output in outputs.items():
-        for irSet in output:
+    for algoName, diag in outputs.items():
+        for irSet in diag.get_output():
             numIRs = irSet.shape[0]
             numSets += 1
     numAlgo = len(outputs)
